pyproject.toml
[tool.poetry]
name = "digletweb"
version = "0.1.0"
description = "Content aggregator and microblogging similar to Digg/Wykop"
authors = ["Kacper Stypik <kacperstypik@mailbox.org>"]
license = "GNU GPL v3"

[tool.poetry.dependencies]
python = "^3.10"
Django = {extras = ["argon2"], version = "^4.1"}
djangorestframework = "^3.13.1"
Pillow = "^9.2.0"
django-environ = "^0.9.0"
whitenoise = "^6.2.0"
psycopg2 = "^2.9.3"
dj-rest-auth = {extras = ["with_social"], version = "^2.2.5"}
djangorestframework-simplejwt = "^5.2.0"
django-model-utils = "^4.2.0"
django-autoslug = "^1.9.8"
Unidecode = "^1.3.4"

[tool.poetry.dev-dependencies]
isort = "^5.10.1"
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
default_section = "THIRDPARTY"
known_first_party = "digletweb"
known_django = "django"
known_drf = "rest_framework"
sections = "FUTURE,STDLIB,DJANGO, DRF, THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
profile = "black"